further development of dht22 interface

+++ main/include/dht22.h
@@ -13,6 +13,6 @@ typedef struct {
 
 esp_err_t dht22_initialize();
 esp_err_t dht22_read(dht22_data * const data);
-uint16_t dht22_get_RH(const dht22_data * const data);
-uint16_t dht22_get_T(const dht22_data * const data);
+float dht22_get_RH(const dht22_data * const data);
+float dht22_get_T(const dht22_data * const data);
 int dht22_is_checksum_valid(const dht22_data * const data);
